Designing Poster, TvShow and Views.
The TvShow organism is created by PosterContent and Poster Image. PosterImage includes the image and PosterContent has the name and user votes.
The View components that were created and Routed were MostPopular, MostViewed, and Most Voted. All of these just show dummie data.
The Routes are in the Content microrganism.